🚚 Move memo arbitrary into /arbitrary
